Change FileObject size from float to long
Updated the SizeInBytes property in FileObject from float to long for more accurate file size representation. Adjusted PrintUtils to cast SizeInBytes to float when calculating percentages.

+++ Source/IO/FileObject.cs
@@ -3,11 +3,11 @@
     public readonly string Name;
     public readonly bool IsDirectory;
     public readonly string Directory;
-    public readonly float SizeInBytes;
+    public readonly long SizeInBytes;
     public readonly bool IsAccessible;
     public readonly List<FileObject> Children;
 
-    public FileObject(string name, bool isDirectory, string directory, float sizeInBytes, bool isAccessible = true)
+    public FileObject(string name, bool isDirectory, string directory, long sizeInBytes, bool isAccessible = true)
     {
         Name = name;
         IsDirectory = isDirectory;
